split route-tree
