+++ shared/actions/chat/creators.js
@@ -4,7 +4,7 @@ import * as I from 'immutable'
 import {chatTab} from '../../constants/tabs'
 import {setRouteState} from '../route-tree'
 
-import type {SetRouteState} from '../../constants/route-tree'
+import type {SetRouteState} from '../../constants/types/route-tree'
 
 function setSelectedRouteState(
   selectedConversation: Types.ConversationIDKey,
